Set a tight file size limit. If the Base64 image gets too big the stylesheet breaks mPDF
A file sized 745kb was about as big as I could get it to keep it working. A switch to binary transfer will make POSTing more efficient, but I'm not sure how to keep the format of the image working. Would either have to consider saving the file locally and referencing the location in the stylesheet, or adding a base64 image an IMG tag in the HTML template. I don't think the PDF will render as reliably.